Separated logic into relevant files
